8. Working with Ref and Forward Ref & Implementation of CartContext, CartProvider and cartReducer to handle Add Item To Cart and Show Number of Cart Items on HeaderCartButton.js using reduce() JS fn.
1 - Working with Ref and Forward Ref:

Here we are using React.forwardRef((props, ref) => {}) and the useRef() Hook to have control on the HTML input inside the Input Component in one hand, and in the other hand, to then retrieve the referenced current value on the Input component in its parent MealItemForm.js.
To export this behavior outside this Input component in MealItemForm.js, we are using :
- the React built-in function forwardRef()
Globally, with forwardRef() we can expose a reference from a targeted HTML element in Input.js to its parent component (MealItemForm.js) to then use the Input component in its parent component through Refs and trigger targeted functionalities.

2 - Implementation of CartContext, CartProvider and cartReducer:

Use of React.createContext({}) and declaration of the Context object in cart-context.js.
Implementation of CartProvider returning the updated cartContext obj (ie. value={cartContext} property) after using the cartReducer which handle dispatching action and updating state.
Instantiation of CartProvider in App.js to make it available to consume for all children.

3 - Handle AddItemToCart and show Nb of Cart Items:

In MealItemForm.js: through props, pass the onAddToCart fn from child to parent with the entered amount number we have retrieved using the useRef hook.
In MealItem.js: with onAddToCart property on the MealItemForm component, pointer to the addToCartHandler in which we call the context(cartCtx) method addItem so we can send the updated obj. Indeed, in CartProvider there is the addItem property available on the cartContext obj pointing to the addItemToCartHandler in which we pass the item obj we sent from MealItem. This Handler will dispatch the action obj to the reducer which in the end will update the existing State.
Finally, in HeaderCartButton.js we also call the useContext(CartContext) Hook so we can display the number of cart items using the reduce() JS fn.