11. Make Items Removable with Reducer
First of all, the CartItem component is used  in CartList.js to display items which comes from the CartProvider.

We attached two handlers to the ADD and REMOVE actions available in the CartItem component (see CartList.js).

In our previous commit, we've managed the addition of items in the CartProvider, now we are handling the removal of items in the CartProvider for which we use a similar reducer logic which consist of:

- to remove an item, we receive in the reducer the item's id that we need to remove or update in the state items list.
The CartItem component is using the onRemove property of its parent CartList.js.
In CartList.js we attached a handler to the remove click action using the bind() js function which allows to preconfigure the parameter of the cartItemRemoveHandler with the item from map ( see bind($arg, item.id))
In the reducer logic, two cases were implemented:
1) if the item's amount is equal to 1, then we need remove it from the state items list (see the if part)
2) if the item's amount is superior to 1, we are decreasing by 1 its amount on the list (see the else part)