20. Move IMG to Firebase Storage
Instead of using local images, we used the Firebase storage service to store our images. To retrieve these img from Firebase we used several functions from SDK Firebase such as ref(), getDownloadURL(), as well as the static method Promise.all() from the Promise API. All of this is possible thanks to our config file generated when we added this web app to our Firebase project. We've also updated some UI Components properties to display correctly images coming from our Firebase storage and our formatted objects in AvailableMenu.js.

Note regarding the last commit:
We wrote a JS script which allow us to generate a JSON file formatted which we can import to Firebase RealTime Database. For the translation, we planned to use the json-translator library for dynamic datas in our script. However we realised that we could face some translations issues due to the Google translate services that is enable by default in the library. We had to use the CLI of this library instead to choose an appropriate provider for our translations.